requirements.txt
fastapi 
strawberry-graphql[fastapi] 
uvicorn
sqlmodel